+++ src/math/noise-generator/noise-generator-gaussian.worker.ts
@@ -11,17 +11,17 @@ self.onmessage = (e) => {
     if (type === MessageToWorker.START) {
         const grid = GridWithMargin.copyWithMargin(data.gridBlueprint);
         const baseGrid = (data.scaleFactor == 1) ? grid : new GridWithoutRange(grid.width, grid.height);
-        let result: Float64Array = calculate(baseGrid, data.mean, data.range, data.standardDeviation);
+        let result: Float32Array = calculate(baseGrid, data.mean, data.range, data.standardDeviation);
         result = upscaleNoise(baseGrid, result, grid, data.scaleFactor);
         console.info(`#NoiseGeneratorGaussian (worker) - calculation done in ${(Date.now() - timeStamp) / 1000}s`);
         self.postMessage({ type: MessageFromWorker.RESULT, result }, [result.buffer]);
     }
 };
 
-function calculate(grid: GridReader, mean: number, range: number, standardDeviation: number): Float64Array {
+function calculate(grid: GridReader, mean: number, range: number, standardDeviation: number): Float32Array {
     const min = mean - range / 2 * standardDeviation;
     const max = mean + range / 2 * standardDeviation;
-    const data = new Float64Array(grid.size);
+    const data = new Float32Array(grid.size);
     for (let row = 0; row < grid.height; row++) {
         for (let col = 0; col < grid.width; col++) {
             let [z0, z1] = boxMullerTransform();
