+++ src/plane/complex-fractal/mandelbrot-combined-iv.ts
@@ -147,7 +147,7 @@ export class MandelbrotCombinedIV extends Plane {
         }
     }
 
-    private async createNoise(sourceGrid: GridWithMargin): Promise<Float64Array> {
+    private async createNoise(sourceGrid: GridWithMargin): Promise<Float32Array> {
         const generator = new NoiseGenerator(sourceGrid);
         const generator$ = generator.createNoise(this.config.data.noiseConfig);
         return await extractData(generator$, 'noise');
